requirements.txt
django
djangorestframework
djangorestframework-simplejwt
djangorestframework-camel-case
factory_boy